# backend/.well-known/openapi.yaml
openapi: 3.0.0
info:
  title: InsightFlow MCP Server API
  description: MCP (Model Context Protocol) server for managing projects, stacks, roadmaps, tasks, and subtasks
  version: 1.0.0
servers:
  - url: http://localhost:3001
    description: Local development server
  - url: https://your-domain.com
    description: Production server
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase authentication token
paths:
...
  /mcp/roadmaps/get:
    post:
      summary: Get roadmap
      description: Get a single roadmap with tasks and subtasks
      operationId: getRoadmap
      tags:
        - Roadmaps
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - roadmap_id
              properties:
                roadmap_id:
                  type: string
      responses:
        '200':
          description: Roadmap details
  /mcp/roadmaps/create:
    post:
      summary: Create roadmap
      description: Create a new roadmap
      operationId: createRoadmap
      tags:
        - Roadmaps
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - project_id
                - name
              properties:
                project_id:
                  type: string
                name:
                  type: string
                description:
                  type: string
                status:
                  type: string
                  enum: [draft, active, completed]
      responses:
        '201':
          description: Roadmap created
  /mcp/roadmaps/update:
    post:
      summary: Update roadmap
      description: Update an existing roadmap
      operationId: updateRoadmap
      tags:
        - Roadmaps
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - roadmap_id
              properties:
                roadmap_id:
                  type: string
                name:
                  type: string
                description:
                  type: string
                status:
                  type: string
      responses:
        '200':
          description: Roadmap updated
  /mcp/roadmaps/delete:
    post:
      summary: Delete roadmap
      description: Delete a roadmap
      operationId: deleteRoadmap
      tags:
        - Roadmaps
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - roadmap_id
              properties:
                roadmap_id:
                  type: string
      responses:
        '200':
          description: Roadmap deleted
...
  /mcp/tasks/get:
    post:
      summary: Get task
      description: Get a single task with subtasks
      operationId: getTask
      tags:
        - Tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - task_id
              properties:
                task_id:
                  type: string
      responses:
        '200':
          description: Task details
  /mcp/tasks/create:
    post:
      summary: Create task
      description: Create a new task
      operationId: createTask
      tags:
        - Tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - roadmap_id
                - name
              properties:
                roadmap_id:
                  type: string
                name:
                  type: string
                description:
                  type: string
                status:
                  type: string
                  enum: [pending, in_progress, completed, blocked]
                priority:
                  type: string
                  enum: [low, medium, high]
                due_date:
                  type: string
      responses:
        '201':
          description: Task created
  /mcp/tasks/update:
    post:
      summary: Update task
      description: Update an existing task
      operationId: updateTask
      tags:
        - Tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - task_id
              properties:
                task_id:
                  type: string
                name:
                  type: string
                description:
                  type: string
                status:
                  type: string
                priority:
                  type: string
                due_date:
                  type: string
      responses:
        '200':
          description: Task updated
  /mcp/tasks/delete:
    post:
      summary: Delete task
      description: Delete a task
      operationId: deleteTask
      tags:
        - Tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - task_id
              properties:
                task_id:
                  type: string
      responses:
        '200':
          description: Task deleted
...